src/pyproject.toml
[project]
name = "cnc-admin"
version = "0.1.0"
description = "Monorepo for CNC admin: API, Worker, and Desktop applications"
requires-python = ">=3.12"

[tool.uv]
package = false
override-dependencies = ["PyQt5-Qt5==5.15.2"]

[tool.uv.workspace]
members = ["core", "api", "worker", "desktop"]

[tool.uv.sources]
cnc-core    = { workspace = true }
cnc-api     = { workspace = true }
cnc-worker  = { workspace = true }
cnc-desktop = { workspace = true }

[dependency-groups]
dev = [
    # Workspace members (installed as editable)
    "cnc-core",
    "cnc-api",
    "cnc-desktop",
    "cnc-worker ; sys_platform == 'linux'",
    # Test tools
    "pytest>=7.4,<8",
    "pytest-cov>=4.1,<5",
    "pytest-env>=1.1,<2",
    "pytest-mock>=3.10,<4",
    "pytest-sqlalchemy-mock",
    "ruff>=0.15.6",
    "ty>=0.0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests", "api/tests", "worker/tests", "desktop/tests"]
pythonpath = ["tests", "api/tests"]
addopts = "--import-mode=importlib --cov=. --cov-report=html"
env = [
    "SERIAL_PORT=COMx",
    "SERIAL_BAUDRATE=115200",
    "USER_ID=1",
]

[tool.coverage.run]
omit = ["**/__init__.py", "tests/*", "*/tests/*", "core/alembic/*", "tests/mocks/*"]

[tool.ruff]
exclude = [
    "venv",
    ".venv",
    "__pycache__",
    "core/alembic/env.py",
    "*.mako",
    "htmlcov",
    "core/core/utilities/grbl/constants.py",
]
line-length = 100

[tool.ruff.lint]
extend-select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "FAST"  # FastAPI
]

[tool.ty]
rules.unresolved-import = "ignore"
src.exclude = [
    "**/conftest.py",
]
